Added the ability to remove bids and recalculate autions
The work for this involves a rewrite of how the auctions are calculated
and how invoices are sent out. Winners are now determined by region and
sent a separate invoice per region. A button was added to the
auction_show partial view to allow an admin to manually kick off an
auction recalculation. This button is only visible if the admin is
viewing the page and the auction is closed.

The recalculation itself runs exactly as a regular auction close but
also checks against the air object when the auction first closed. It
checks to see if there are any existing winners and what the difference
in slots are like for each existing winner. If there's no existing
winner, the new winner is treated like any first time winner.

The auction_show partial view was also modified so that when an admin is
viewing the page, each slot has a control to delete the bid associated
with the slot. The backend code already existed for this so just the
route itself need to be wired up.