Invoices for Auction view
Created two views that allow an admin to view all the invoices for an
auction. The first view allows the admin to choose an auction to view
invoices for. The second view allow the admin to view all the invoices
for an auction.

The work for this included creating a new couchdb view, a new db method
and a new model and route file. This also includes updating the site,
model and router files to account for the new items.

Also added a link to the auction invoice views from the admin page. The
link is located in the nav bar for the admin page.