Invalidate bid by admin changed and recalc fixes
Before, when an admin removed a bid from the auction post auction
closing, all it did was mark the bid as invalid. Now, all bids related
to the user for the given bid's region in an auction are marked invalid
and the invoices that are attached to those bids (if they exist) are
void. This only tries to void those invoices as invoices already paid or
in the process of being paid are not able to be void. Also an email is
sent to the bidder to notify them that their bids are now invalid.

To accomplish this, a new couchapp view was needed to return all the
invoice for a bidder in an auction. From there, all the invoices with
a matching region to the original invalid bid are also invalidated.

This required a new DB method to utilize the couchapp view and a few
extra functions in the bid route to invalidate the correct bids and
invoices, then send the bidder an email. All objects updated are sent to
the auction view for the admin to see what changed and to the bidder in
the email they receive.

Fixes:
Recalculations were previously voiding all bids in all regions of an
auction if an invoice was not paid, now only the bids in the region
where the invoice was not paid are being voided. This required updating
the couchapp function 'getUserBidsPerAuction' to 'getUserBidsPerRegion'
to return all bids by a user in an auction's region rather than all bids
by a user in an auction entirely. The DB method that was tied to
'getUserBidsPerAuction' was also updated and renamed. It also required
update the transient bid object in the recalculation to also region the
bid is for, for later comparison during bid invalidation.