Added Reserved Ads
Created a reservedAd model that represents an ad that is only created by
an admin and is automatically injected into all auction regions. This
allows for a universal ad that is maintained by the admins.

The first step was to create a DB model that represented a reserved ad.
This ad is similar to any other ad except that the only modifiers it has
is an in_use property that the admin can use to put a reserved ad into
use or not. Otherwise, it maintains the same HTML and CSS fields. It
also contains the user that created this ad and the regions that this ad
applies to.

A couchDB view was created to return all reserved ads as well.

DB methods were created to get a single reserved ad by id, get all
reserved ads, create and  update a reserved ad and finally delete a
reserved ad by id.

Two new models were added to emcee. One is to retrieve a single reserved
ad by id and another to retrieve all reserved ads.

New routes were added to both the site.js file and the router.js file to
serve the reservedAds from the /admin/ads/reserve path.

A new view was added to show all the reservedAds and a new version of
the ad_upload view was created to edit reservedAds.

Along with the code to represent and save reservedAds, the
auction_probability.js file and the auction_show and index_show views
were edited to properly show the probabilities for each region of an
auction. This takes into account the slots for that particular region,
the matching reserved ads for that region and plus one for the factoid
slot that is still to be written.