Bug Fixes
Added another check to the parseBool function in config. Checked that
the incoming value is not undefined and if it is, return false.

Added the admin check to all the routes in the route file,
reserved_ads.js. Before it would allow any logged in user to
add/update/delete reservedAds. Now only admin can fire off those
actions.

Rewrote the conditional logic for the model file, reservedAds.js.
Originally if was using an if(!err) which is more tedious to read
compared to if (err) which it was been changed to.