Added a dbVersion check
index.js now contains and runs a check to ensure that any changes to the
couchapp is accounted for by a variable called dbVersion.

Any changes to couchapp at this point forward should update the
hardcoded dbVersion var before committing.

If the couchapp file is updated and the couch that this code points to
is not updated. The startup process is stop at that point and will not
start adness until the version is updated in the db.