Config.js parseNumber method
Added a method to the config.js file that allows the env variables to be
parsed as numbers but unlike the usual Number(var), the parseNumber
method will return zero without being evaluated to false. The previous
method of doing process.env.VAR_NAME || defaultVAR would evaluate to the
defaultVAR if the VAR_NAME was set to the number 0.