Modified Registration Fee to be a one time discount
Modified the Registration Fee to be a discount on the user's first
invoice after paying the registration fee. This required editing the
invoice.js file to account for the way Baron handle discounts. It also
required changes to auction-close.js to check the registeredUser and see
if the user has any discount remaining. And to also clear out the
discount when it's used.

The db.js file was edited to allow the discount_remaining property to be
allowed through when saving/editing the registered user.

Finally the webhook.js file needed to be edited so actually enter the
discount amount into the registered user when they paid their
registration fee.