Updated Edit Ad Edit Button to be admin only
Updated the floating Edit Button in the ad/show partial view so that
it's only viewable by admins.

Updated the config.fakeauth.admin's default to false since parseBool can
return false. At this point, the the default is used since the OR
operator takes in the false return by parseBool as a signal to use the
default.