Bug Fixes
Fixed a circular dependency issue with auction-close and invoice. Fixed
this by requiring the invoice.js module only when needed inside the
function that generates invoices in auction-close.

Also fixed a race issue within the same function. RegisteredUser was
being updated twice when the user has a discount. Now the branch that
handles discounts updates the registeredUser and not at a higher level.