Fixed a bug with floating point numbers on bids
Fixed a bug where the bid price could be represented as a long floating
point value. Used .toFixed(2) to ensure the bid price input value is
never longer than 2 decimal places. This function is in place in both
the view and in the route file to ensure it's never represented as
anything other a number with two decimal places.

closes #62