Removed unregistered user's ads from admin' ad listing
Removed unregisterd user's ads from the 'Needs review', 'Approved Ads',
and 'Rejected Ads' views. To ensure that resources were used optimally
the linked document feature of couchDB was used to return all the
submitted/approved/rejected ads along with their respective user. To be
able to do this, the registeredUser DB model was edited so that the _id
for registeredUser was replaced with the userId.

To keep backwards compatibility, the webhook.js, profileUser.js,
profile.js,and the registration.js files were updated to reference the
registeredUser object by it's new _id. Updates were also made so that
when a new registeredUser object was created, the correct _id was used.
The db.js file was also updated to utilize the new linked document view
correctly since the values are now located in the 'doc' property of the
return body instead of in the value property.

Finally the approvedAds.js, rejectedAds.js and the reviewAds.js files
were also updated to separate the registeredUser objects from the ad
objects. For each ad object the respective user's object was checked to
see if the user was registered. If they weren't, the ad was not
displayed.

Also added an alert to all profile that have an ad, a signed in user's
page, or any user's page that has registered but have not paid the fee.
This alert basically tells the user that until they've paid their
registration fee, the ads created on that account will not be viewable
by the admins.

closes #61