Added bid validation
Created a model that returns all the bids for a user in a given auction
region. This model is used in the newBid method of routes/bid.js to
check the highest price per bid and to ensure that any new bids are
higher than that price.

The bid is also validated to be equal to the minimum bid increment step
of 0.05 (hardcoded). But at the same time are not validated against the
winning bid threshold for now. If any of these validations do not pass,
the bid is ntot saved and the errors are sent to the user.

The db errors for newBids were re-written to be more node compliant and
when returned to the newBid route, those errors are also displayed to
the user using connect flash. Being so, auction.js needed to be
modified to pass the errors into the page. Auction_show.ejs was also
edited to show those errors. These errors include all the validation
errors.

closes #68
closes #70
closes #71