New bids replace all previous bids and bug fixes
Adjusted the bidding algorithm so that there is every only one bid that
actually matters in each auction region for each user. Any new bids that
pass validation will replace the old highest bid.

The fix itself is implemented in the appendBidsToAuction function in
db.js. For each Auction Region, after all the invalid bids are removed
and the bids for only that region are pulled out, the highest bid for
each user is pulled aside and only those bids are sent to the bidding
algorithm. All other bids are considered valid but unused for now.

Bug Fix:
Fixed a bug with routes/bid.js where the minimum increment validation
doesn't always work and throws false positives. The problem mainly stems
from the fact that Javascript can't handle decimal modulus very well.
The fix is to multiply both the minimum increment and the incoming bid
price by 100 and then run the modulus. 100 is a reasonable number since
we're only working up to that precision.

closes #75