User Suspension and bug fixes
Added an extra validation to the routes/bid.js file in the newBid
function so that a user that has been manually suspended can no longer
bid on any auction.

Bug Fixes:
Also added another validation so that a user that has not registered
cannot bid on any auction as well. There is another validation for this
exact reason in the newBid function in db.js and that validation will
stay there for now until a good structure for this logic is figured out.

registration.js - Capitalization and wording fix for the invoice
function since the only possible route for this function to be called is
through the routes/registration.js file. User is this case will always
be the logged in user.