Admin Delete Bid only affects bid in question
Before, when an admin deleted a bid from the auction view, it would
invalidate all bids for that user in that region as well as voiding all
their invoices. Now, the same button has been changed to just invalidate
the bid that was clicked on and only void invoices that pertain to that
bid.

closes #69