Split bid history by region
Split the bid history from one large section with all bids sorted by
creation time into a section for each region but still sorted by
creation time. The sorting lists the bid in decending order so the first
bid is on top and the last bid on the bottom.

The routes/common/auction.js file was edited to arrange the bids into
sections and then sorts and update the creation time to something human
readable. The auction.ejs file was edited to match the
routes/common/auction.js file.

closes #76