HTML entities encoded in ad editor
This is a fix for issue #84 where HTML Codes were converted and saved as
their converted form into the DB and not as their escaped codes.

This still creates an issue after the code is saved to the DB. On any
future edits, the code is converted in the textarea and could possibly
be saved into the DB as its converted form at that point.