Bug Fixes for factoids and showing HTML codes in editor
Fixed an issue with HTML Codes not showing up in the editor properly.
Now the editor shows the HTML codes instead of the converted value.

Fixed a issue with factoids where clicking on the factoid row deleted
the factoid. Now only clicking on the red X will delete the factoid.
Also factoids can now contain HTML and the HTML is validated as well.

These fixes also required that the replacement text in the factoid main
HTML be changed from <%- text %> to &lt;%- text %&gt; mainly to handle
the issue with HTML codes not showing up properly.

closes #84